feat(ssh): add pub keys to known hosts
This is primarily used to ensure that we can connect to ssh targets
without needing to disable StrictHostKeyChecking. Storing the public
keys in the role/group_vars mean that we can ensure that the keys won't
change, even between deployments of these playbooks.

If a key isn't provided for any reason, it can be determined at
runtime by doing an `ssh-keyscan`, although this is not ideal.