refactor(git): smarter handling of git user details
We're taking the git user details directly when calling the role. This
allows us to use the variables we're loading dynamically from
.playbooks.yml.

We're also storing em as a dict; it'll be easier to suppress the user
section if no details are provided but it'll also be easier to add other
fields to that section.