feat(fluentd): @include config from other sources
Creating a conf.d/ allows us to maintain a default config with important
system-wide settings (currently nothing) and also provide a mechanism
for admins to write additional config to integrate systems and have it
dropped in during a playbook run.

To include custom config, include files in playbook's template folder
and ensure the path is covered by the fluentd_config_paths variable.